splitting the todos
