+++ flutter_mobx/example/lib/todos/todo.dart
@@ -1,0 +1,15 @@
+import 'package:mobx/mobx.dart';
+
+part 'todo.g.dart';
+
+class Todo = TodoBase with _$Todo;
+
+abstract class TodoBase implements Store {
+  TodoBase(this.description);
+
+  @observable
+  String description = '';
+
+  @observable
+  bool done = false;
+}
